requirements.txt
opencv-python
numpy
Pillow
matplotlib
scikit-image
